labs/capstone-options/option-a-code-review/backend/config.py
"""Application configuration from environment variables."""
import os
from dotenv import load_dotenv

load_dotenv()

ANTHROPIC_API_KEY = os.getenv("ANTHROPIC_API_KEY", "")
ANTHROPIC_MODEL = os.getenv("ANTHROPIC_MODEL", "claude-sonnet-4-20250514")

GITHUB_WEBHOOK_SECRET = os.getenv("GITHUB_WEBHOOK_SECRET", "")
GITHUB_TOKEN = os.getenv("GITHUB_TOKEN", "")

ALLOWED_ORIGINS = os.getenv(
    "ALLOWED_ORIGINS",
    "http://localhost:3000"
).split(",")

RATE_LIMIT = os.getenv("RATE_LIMIT", "10/minute")
MAX_CODE_SIZE = int(os.getenv("MAX_CODE_SIZE", "51200"))  # 50KB
