labs/capstone-options/option-a-code-review/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.10.0
anthropic>=0.40.0
python-dotenv==1.0.1
slowapi==0.1.9
httpx==0.27.0
